Stop cover between eunit runs for speed.
Cover gets slower and slower for each application. This is due to
the cover_server internal state. Stopping the cover between eunit+cover
runs, emptying the cover_server state, gives a ~5-6 speed improvement
on systems with a lot of dependencies. Stopping the cover replaces the
earlier practice of doing a cover:reset before each run.